add errors to severity functions
The functions dealing with severities did not properly use error codes.
This change updates these functions to generate appropriate error messages
when a failure occurs, so that users checking for errors are properly
notified. Errors are also cleared in the event of success so that an
error is not mistakenly identified for a successful call.

The affected functions are:
  stumpless_get_severity_string
  stumpless_get_severity_enum
  stumpless_get_severity_enum_from_buffer