update and refactor cmake scripts
The CMake minimum requirement of 3.2 was too low for recent versions of
CMake, which caused builds to fail. In addition, several features from
newer versions of CMake allow the build scripts to be more resilient and
efficient. This refactoring raises the minimum requirement to 3.21,
supporting several improvements to the build process along the way.

The dynamic download of Google Test and Benchmark with ExternalProject has
been replaced with FetchContent paired with FIND_PACKAGE_ARGS to find a
local installation and fall back on a download and build. This necessitated
the removal of the export-gtest and export-benchmark build targets and the
GTEST_PATH and BENCHMARK_PATH build options.

The building of testing and benchmarking can be controlled with the new
BUILD_TESTING and BUILD_BENCHMARKING build options. A C++ compiler is no
longer required to build the project if both of these are disabled.

Many of the CMake source files have been refactored to more explicitly
provide full paths to files, use quotes to designate strings where they
are intended, and break apart functionality into modular files.